+++ users/apps.py
@@ -3,3 +3,6 @@ from django.apps import AppConfig
 
 class UsersConfig(AppConfig):
     name = "users"
+
+    def ready(self):
+        import users.signals  # noqa
